Move hook listener classes to EventListener\Hook namespace.
