+++ src/EventListener/Hook/RateItArticleListener.php
@@ -14,13 +14,14 @@
  * @filesource
  */
 
-namespace Hofff\Contao\RateIt;
+namespace Hofff\Contao\RateIt\EventListener\Hook;
 
 use Contao\File;
 use Contao\FilesModel;
 use Contao\StringUtil;
+use Hofff\Contao\RateIt\RateItFrontend;
 
-class RateItArticle extends RateItFrontend
+class RateItArticleListener extends RateItFrontend
 {
 
     /**
