+++ src/EventListener/Hook/RateItPageListener.php
@@ -14,11 +14,12 @@
  * @filesource
  */
 
-namespace Hofff\Contao\RateIt;
+namespace Hofff\Contao\RateIt\EventListener\Hook;
 
 use Contao\Frontend;
+use Hofff\Contao\RateIt\RateItRating;
 
-class RateItPage extends Frontend
+class RateItPageListener extends Frontend
 {
 
     var $rateItRating;
